Update R/W sweep experiments and scripts

+++ scripts/rw_sweep/run.sh
@@ -5,46 +5,89 @@ script_loc=$(cd $(dirname $0) && pwd -P)
 cd $script_loc
 source ../experiment_config.sh
 
-if [ -z $1 ]; then
-  >&2 echo "Usage: $0 <checkpoint name> [other args passed to run_custom]"
+# Evaluates any environment variables in this script's arguments. This script
+# should only be run on trusted input.
+orig_args=($@)
+all_args=()
+for val in "${orig_args[@]}"; do
+  phys_arg=$(eval "echo $val")
+
+  # Extract the database type (e.g., llsm, rocksdb)
+  if [[ $phys_arg =~ --db=.+ ]]; then
+    db_type=${phys_arg:5}
+  fi
+
+  # Extract the checkpoint name, which shouldn't be passed as an argument further.
+  # Add anything else to args.
+  if [[ $phys_arg =~ --checkpoint_name=.+ ]]; then
+    checkpoint_name=${phys_arg:18}
+  else
+    all_args+=($phys_arg)
+  fi
+done
+
+if [[ -z $checkpoint_name ]]; then
+  echo >&2 "Usage: $0 --checkpoint_name=<checkpoint name> [other args passed to run_custom]"
   exit 1
 fi
 
-checkpoint_name=$1
-shift 1
-
-# Evaluates any environment variables in this script's arguments. This script
-# should only be run on trusted input.
-orig_args=$@
-all_args=()
-for val in "${orig_args[@]}"; do
-  all_args+=$(eval "echo $val")
-done
-
 # Generates the workload configuration and filters out unrelated arguments.
-args=$(python3 generate_workload.py $all_args)
+args=$(python3 generate_workload.py ${all_args[@]})
 
+echo "Detected DB Type: $db_type"
+echo "Detected checkpoint name: $checkpoint_name"
+
+# Add additional common arguments.
+args+=("--verbose")
+args+=("--db_path=$DB_PATH")
+args+=("--seed=$SEED")
+args+=("--notify_after_init")
+args+=("--skip_load")
+
+# Use the generated workload file.
+args+=("--workload_config=$COND_OUT/workload.yml")
+
+# Copy over the checkpoint.
 full_checkpoint_path=$DB_CHECKPOINT_PATH/$checkpoint_name
-
 rm -rf $DB_PATH
 cp -r $full_checkpoint_path $DB_PATH
 sync $DB_PATH
 
+init_finished=0
+
+function on_init_finish() {
+  init_finished=1
+}
+
+# Interrupt the first `wait` below when we receive a `SIGUSR1` signal.
+trap "on_init_finish" USR1
+
 set +e
-iostat -o JSON -d -y 1 > $COND_OUT/iostat.json &
+../../build/bench/run_custom ${args[@]} >$COND_OUT/results.csv &
+wait %1
+code=$?
+
+# The experiment failed before it finished initialization.
+if [ "$init_finished" -eq "0" ]; then
+  exit $code
+fi
+
+# The DB has finished initializing, so start `iostat`.
+iostat -o JSON -d -y 1 >$COND_OUT/iostat.json &
 iostat_pid=$!
 
-../../build/bench/run_custom \
-  --verbose \
-  --db_path=$DB_PATH \
-  --seed=$SEED \
-  --skip_load \
-  --workload_config=$COND_OUT/workload.yml \
-  $args \
-  > $COND_OUT/results.csv
+# Wait until the workload completes.
+wait %1
+code=$?
 
-cp $DB_PATH/llsm/LOG $COND_OUT/llsm.log
-cp $DB_PATH/rocksdb/LOG $COND_OUT/rocksdb.log
-
+# Stop `iostat`.
 kill -s SIGINT -- $iostat_pid
 wait
+
+cp $DB_PATH/$db_type/LOG $COND_OUT/$db_type.log
+du -b $DB_PATH >$COND_OUT/db_space.log
+
+# Report that the experiment failed if the `run_custom` exit code is not 0
+if [ $code -ne 0 ]; then
+  exit $code
+fi
